requirements.txt
python-telegram-bot==13.7
openai
python-dotenv
gspread
oauth2client